backend/requirements.txt
Flask
flask-cors
requests
beautifulsoup4
chromadb
langchain
tiktoken
openai
python-dotenv
